properly check for the existance of the session

+++ django_shared_auth/middleware.py
@@ -69,8 +69,10 @@ class SharedAuthProviderMiddleware(object):
         the at cookie should be deleted.
         """
         modified = True
-        if request.session and request.user.is_authenticated() and \
-            not hasattr(request, settings.COOKIE_NAME):
+        if getattr(request, 'session', None) and \
+                hasattr(request, 'user') and \
+                request.user.is_authenticated() and \
+                not hasattr(request, settings.COOKIE_NAME):
 
             if request.session.get_expire_at_browser_close():
                 max_age = None
@@ -86,8 +88,10 @@ class SharedAuthProviderMiddleware(object):
                     domain=settings.COOKIE_DOMAIN,
                     path=settings.COOKIE_PATH,
                     secure=settings.SECURE)
-        if not (request.session and request.user.is_authenticated()) \
-            and hasattr(request, settings.COOKIE_NAME):
+        if not (getattr(request, 'session', None)
+                and hasattr(request, 'user')
+                and request.user.is_authenticated()) \
+                and hasattr(request, settings.COOKIE_NAME):
             response.delete_cookie(settings.COOKIE_NAME)
         return response
